# core/docs/swagger.yaml
definitions:
...
  define.ChangeHostGroup:
    properties:
      addrs:
        description: 主机host
        items:
          type: string
        type: array
      id:
        type: string
      remark:
        type: string
    required:
    - id
    type: object
...
  define.CreateHostGroup:
    properties:
      addrs:
        description: 主机host
        items:
          type: string
        type: array
      name:
        type: string
      remark:
        type: string
    required:
    - name
    type: object
  define.GetID:
    properties:
      id:
        type: string
    type: object
...
  resp.Response:
    properties:
      code:
        type: integer
      data:
        type: object
      msg:
        type: string
    type: object
info:
  contact:
    email: labulakalia@gmail.com
    name: labulaka521
    url: http://www.swagger.io/support
  description: Crocodile Swaager JSON API
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: https://github.com/labulaka521/crocodile
  title: Crocidle API
  version: "1.0"
paths:
  /api/v1/host:
    delete:
      description: delete host
      parameters:
      - description: ID
        in: body
        name: StopHost
        required: true
        schema:
          $ref: '#/definitions/define.GetID'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.Response'
      security:
      - ApiKeyAuth: []
      summary: delete host
      tags:
      - Host
    get:
      description: get all registry host
      parameters:
      - description: Offset
        in: query
        name: offset
        type: integer
      - description: Limit
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.Response'
      security:
      - ApiKeyAuth: []
      summary: get all hosts
      tags:
      - Host
  /api/v1/host/stop:
    put:
      description: stop host worker
      parameters:
      - description: ID
        in: body
        name: StopHost
        required: true
        schema:
          $ref: '#/definitions/define.GetID'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.Response'
      security:
      - ApiKeyAuth: []
      summary: stop host worker
      tags:
      - Host
  /api/v1/hostgroup:
    delete:
      description: delete hostgroup
      parameters:
      - description: HostGroup
        in: body
        name: User
        required: true
        schema:
          $ref: '#/definitions/define.GetID'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.Response'
      security:
      - ApiKeyAuth: []
      summary: delete hostgroup
      tags:
      - HostGroup
    get:
      description: get all hostgroup
      parameters:
      - description: Offset
        in: query
        name: offset
        type: integer
      - description: Limit
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.Response'
      security:
      - ApiKeyAuth: []
      summary: get all hostgroup
      tags:
      - HostGroup
    post:
      description: create new hostgroup
      parameters:
      - description: HostGroup
        in: body
        name: User
        required: true
        schema:
          $ref: '#/definitions/define.CreateHostGroup'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.Response'
      security:
      - ApiKeyAuth: []
      summary: create hostgroup
      tags:
      - HostGroup
    put:
      description: change hostgroup
      parameters:
      - description: HostGroup
        in: body
        name: User
        required: true
        schema:
          $ref: '#/definitions/define.ChangeHostGroup'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.Response'
      security:
      - ApiKeyAuth: []
      summary: change hostgroup
      tags:
      - HostGroup
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
  BasicAuth:
    type: basic
swagger: "2.0"